backend/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
sqlalchemy
alembic
python-multipart
pymysql
python-jose[cryptography]
passlib[bcrypt]
openpyxl
